Added waiting for keyboard to be found loss
cleaned up some misc code all over
fixed all of the warnings

+++ DotNetK70Controller/Config.h
@@ -39,7 +39,7 @@ void SetupConfig()
 
 			bool EqualsFound = false;
 
-			int i = 0;
+			size_t i = 0;
 			while (!Done) // parses the current line
 			{	
 				char CurChar = CurLine.at(i);
